+++ dbdaora/geospatial/_tests/mongodb/test_integration_service_geospatial_aioredis_mongodb_delete.py
@@ -1,7 +1,6 @@
 import asynctest
 import pytest
 from aioredis import RedisError
-from jsondaora import dataclasses
 
 from dbdaora.exceptions import EntityNotFoundError
 
@@ -10,44 +9,57 @@ from dbdaora.exceptions import EntityNotFoundError
 async def test_should_delete(
     fake_service, serialized_fake_entity, fake_entity
 ):
+    await fake_service.repository.memory_data_source.delete('fake:fake2:fake')
     await fake_service.add(fake_entity)
 
-    assert await fake_service.get_one('fake', other_id='other_fake')
+    assert await fake_service.get_one(
+        fake_id=fake_entity.fake_id,
+        fake2_id=fake_entity.fake2_id,
+        latitude=5,
+        longitude=6,
+        max_distance=1,
+    )
 
-    await fake_service.delete(fake_entity.id, other_id='other_fake')
-    fake_service.cache.clear()
+    await fake_service.delete(
+        fake_id=fake_entity.fake_id, fake2_id=fake_entity.fake2_id,
+    )
 
     with pytest.raises(EntityNotFoundError):
-        await fake_service.get_one('fake', other_id='other_fake')
+        await fake_service.get_one(
+            fake_id=fake_entity.fake_id,
+            fake2_id=fake_entity.fake2_id,
+            latitude=5,
+            longitude=6,
+            max_distance=1,
+        )
 
 
 @pytest.mark.asyncio
 async def test_should_delete_from_fallback_after_open_circuit_breaker(
-    fake_service, serialized_fake_entity, fake_entity, mocker
+    fake_service, fake_entity, fake_fallback_data_entity, repository
 ):
-    await fake_service.repository.memory_data_source.delete(
-        'fake:other_fake:fake'
-    )
-    await fake_service.repository.memory_data_source.delete(
-        'fake:not-found:other_fake:fake'
-    )
+    await fake_service.repository.memory_data_source.delete('fake:fake2:fake')
     key = fake_service.repository.fallback_data_source.make_key(
-        'fake', 'other_fake', 'fake'
+        'fake', 'fake2', 'fake'
     )
     await fake_service.repository.fallback_data_source.put(
-        key, dataclasses.asdict(fake_entity, dumps_value=True)
+        key, fake_fallback_data_entity
     )
 
-    assert await fake_service.get_one('fake', other_id='other_fake')
+    assert await fake_service.get_one(
+        fake_id=fake_entity.fake_id,
+        fake2_id=fake_entity.fake2_id,
+        latitude=5,
+        longitude=6,
+        max_distance=1,
+    )
 
     fake_service.repository.memory_data_source.delete = asynctest.CoroutineMock(
         side_effect=RedisError
     )
 
-    await fake_service.delete(fake_entity.id, other_id='other_fake')
-    fake_service.cache.clear()
+    await fake_service.delete(
+        fake_id=fake_entity.fake_id, fake2_id=fake_entity.fake2_id,
+    )
 
-    with pytest.raises(EntityNotFoundError):
-        await fake_service.get_one('fake', other_id='other_fake')
-
-    assert fake_service.logger.warning.call_count == 2
+    assert fake_service.logger.warning.call_count == 1